requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aiohttp==3.9.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.16
langchain-community==0.0.33
    # via langchain
langchain-core==0.1.45
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
    # via langchain
langchainhub==0.1.15
langsmith==0.1.49
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.21.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.23.1
    # via langchain-openai
orjson==3.10.1
    # via langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pydantic==2.7.0
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.18.1
    # via pydantic
python-dotenv==1.0.1
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.4.16
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langchain-community
    #   langchainhub
    #   langsmith
    #   tiktoken
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.29
    # via
    #   langchain
    #   langchain-community
sympy==1.12
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.6.0
    # via langchain-openai
tqdm==4.66.2
    # via openai
types-requests==2.31.0.20240406
    # via langchainhub
typing-extensions==4.11.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.1
    # via
    #   requests
    #   types-requests
yarl==1.9.4
    # via aiohttp
